move survey helper and test helpers to prompt pkg
