+++ pkg/prompt/prompt.go
@@ -20,3 +20,7 @@ var Confirm = func(prompt string, result *bool) error {
 	}
 	return survey.AskOne(p, result)
 }
+
+var SurveyAsk = func(qs []*survey.Question, response interface{}, opts ...survey.AskOpt) error {
+	return survey.Ask(qs, response, opts...)
+}
